Rename Assets::Source to just Assets
Assets::Source always felt awkward, especially given we had to pass `source_path` arguments to its methods. In the end, I couldn’t think of any better name for this class (whose role is to provide access to the assets themselves) than just straight up `Assets`. Usage now feels much more natural.
